+++ src/pico.h
@@ -430,6 +430,11 @@ Pico_Color  pico_get_pencil_color (void);
 const char* pico_get_pencil_font  (void);
 PICO_STYLE  pico_get_pencil_style (void);
 
+/// @brief Gets the state of auto aids.
+/// @return 1 if enabled, or 0 otherwise
+/// @sa pico_set_aids
+int pico_get_aids (void);
+
 /// @brief Gets the state of expert mode.
 /// @param fps optional pointer to receive fps value (NULL to ignore)
 /// @return 1 if enabled, or 0 otherwise
@@ -687,6 +692,11 @@ void pico_set_pencil       (Pico_Layer_Pencil pencil);
 void pico_set_pencil_color (Pico_Color color);
 void pico_set_pencil_font  (const char* path);
 void pico_set_pencil_style (PICO_STYLE style);
+
+/// @brief Toggles auto aids (quit/exit, window resize, ctrl+keys).
+/// @param on 1 to enable, 0 to disable
+/// @sa pico_get_aids
+void pico_set_aids (int on);
 
 /// @brief Toggles expert mode (manual present, frame pacing).
 /// @param on 1 to enable, 0 to disable